client-py/requirements.txt
grpcio==1.65.1
grpcio-tools==1.65.1